Numeric arguments can accept a double in tht check.

+++ src/engine/CSVCheckEngine.java
@@ -112,7 +112,7 @@ public class CSVCheckEngine {
 			
 			try{
 				
-				Integer.parseInt(argument);
+				Double.parseDouble(argument);
 				
 				isArgument = true;
